Move SettingValue types to own file
